% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{coef.matchpropexc}
\alias{coef.matchpropexc}
\title{coefficient estimates}
\usage{
\method{coef}{matchpropexc}(object, ...)
}
\arguments{
\item{object}{model estimated with matchpropexc}
}
\description{
coefficient estimates
}
\author{
Cristina Boschini
}
