% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{compdata}
\alias{compdata}
\title{Data structured for matchpropexc}
\usage{
compdata(formula, data, clust, idControl, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph}); avoid using strata() or factor()}

\item{data}{data frame}

\item{clust}{vector exposed indicator; it identifies the "cluster" (one exposed in each clust i)}

\item{idControl}{vector control indicator (idControl==1 indicates exposed individual in clust i)}

\item{...}{Additional arguments to lower level funtions}
}
\value{
A setup dataset, ready for \code{matchpropexc}
}
\description{
Data structured for matchpropexc
}
\examples{
dd<-data.sim(nca=5000, ncont=5)
setdd<-compdata(Surv(time, status)~x+z+cc, clust=id, idControl=j, data=dd)
summary(setdd)
}
\author{
Cristina Boschini
}
