% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{cumhaz.matchf}
\alias{cumhaz.matchf}
\title{Cumulative baseline excess hazard estimate}
\usage{
cumhaz.matchf(object, strata = object$strata, time = NULL,
  SEcumhaz = TRUE)
}
\arguments{
\item{object}{model estimated with matchpropexc}

\item{strata}{default strata if the model has them.}

\item{time}{specify at which time to compute the estimates}

\item{SEcumhaz}{by default set to TRUE. FALSE if you don't want to compute it.}
}
\description{
Cumulative baseline excess hazard estimate
}
\author{
Cristina Boschini
}
