% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSim.R
\name{data.sim}
\alias{data.sim}
\title{Function to simulate macthed survival data}
\usage{
data.sim(nca, ncont, competing = FALSE, nullmod = FALSE)
}
\arguments{
\item{nca}{number of exposed individuals}

\item{ncont}{number of unexposed individuals for eac exposed (fixed number)}

\item{competing}{if TRUE a competing cause is considered}

\item{nullmod}{if TRUE no covariates are simulated. By default 3 variables are simulated: two binomals and one continuous.}
}
\description{
Function to simulate macthed survival data
}
\author{
Cristina Boschini
}
