% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchcif.R
\name{ecif.coef}
\alias{ecif.coef}
\title{Excess CIF prediction based on newdata and geepack::geese estimates.}
\usage{
ecif.coef(model, times, link = "log")
}
\arguments{
\item{model}{geese object. To be defined if coefs and vcov are null.}

\item{times}{vector of timepoints as the one used to estimate the GEE model}

\item{link}{link used to estimate the model. Choose between c("log","logit","identity")}

\item{labels}{vector of parameter names}
}
\value{
table with coefficient estimates, sandwich standard error, function of coefficient for interpretation and p-value.
}
\description{
Excess CIF prediction based on newdata and geepack::geese estimates.
}
