% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{ehaz.plot}
\alias{ehaz.plot}
\title{Baseline cumulative excess estimates arranged to be plotted with ggplot2}
\usage{
ehaz.plot(object, time = NULL, relsurv = FALSE, level = 0.95)
}
\arguments{
\item{object}{model estimated with matchpropexc}

\item{time}{if baseline hazed should be visualized only at specific time points}

\item{relsurv}{if relative survival quantities should be computed}

\item{level}{CI level}
}
\value{
data table to be easily used in ggplot2
}
\description{
Baseline cumulative excess estimates arranged to be plotted with ggplot2
}
\author{
Cristina Boschini
}
