% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{exccumhaz}
\alias{exccumhaz}
\title{Cumulative baseline excess hazard estimate}
\usage{
exccumhaz(object, strata = object$strata, time = NULL,
  SEcumhaz = TRUE)
}
\arguments{
\item{object}{model estimated with matchpropexc}

\item{strata}{default strata if the model has them.}

\item{time}{specify at which time to compute the estimates}

\item{SEcumhaz}{by default set to TRUE. FALSE if you don't want to compute it.}
}
\value{
cumulative baseline excess hazard estimates. If the model has strata, the returned object will be a list. Estimated are computed at the defined time.
}
\description{
Cumulative baseline excess hazard estimate
}
\examples{
dhaz<-sim.data.MatchH(nca=5000, ncont=5)
setdhaz<-compdata(Surv(time, status)~x+z+cc, clust=id, idControl=j, data=dhaz)
exc.model<-matchpropexc(Surv(entry,exit,status)~strata(z)+factor(x), data=setdhaz)
cumhaz <- exccumhaz(exc.model) #it's a list because of strata
cumhaz <- exccumhaz(exc.model, time=seq(0,30,5)) 
#you can chose at which time-points to show the estimates
}
\author{
Cristina Boschini
}
