% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{excplot.matchpropexc}
\alias{excplot.matchpropexc}
\title{Plotting the cumulative baseline excess hazard}
\usage{
excplot.matchpropexc(x, se = FALSE, time = NULL, add = FALSE,
  ylim = NULL, lty = NULL, col = NULL, legend = TRUE,
  ylab = "Baseline cumulative excess hazard", polygon = TRUE,
  level = 0.95, stratas = NULL, relsurv = FALSE, ...)
}
\arguments{
\item{x}{matchpropexc object}

\item{se}{to include standard errors}

\item{time}{to plot for specific time variables}

\item{add}{to add to previous plot}

\item{ylim}{to give ylim}

\item{lty}{to specify lty of components}

\item{col}{to specify col of components}

\item{legend}{to specify col of components}

\item{ylab}{to specify ylab}

\item{polygon}{to get standard error in shaded form}

\item{level}{of standard errors}

\item{stratas}{wich strata to plot (number or vector between 0 and nstrata-1)}

\item{relsurv}{set to FALSE. If TRUE relative survival curves are plotted.}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Plotting the cumulative baselines excess hazard
}
\author{
Cristina
}
