% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{excplot}
\alias{excplot}
\title{Plotting the cumulative baseline excess hazard}
\usage{
excplot(x, se = FALSE, time = NULL, add = FALSE, ylim = NULL,
  lty = NULL, col = NULL, legend = TRUE,
  ylab = "Baseline cumulative excess hazard", polygon = TRUE,
  level = 0.95, stratas = NULL, relsurv = FALSE, ...)
}
\arguments{
\item{x}{matchpropexc object}

\item{se}{to include standard errors}

\item{time}{to plot for specific time variables}

\item{add}{to add to previous plot}

\item{ylim}{to give ylim}

\item{lty}{to specify lty of components}

\item{col}{to specify col of components}

\item{legend}{to specify col of components}

\item{ylab}{to specify ylab}

\item{polygon}{to get standard error in shaded form}

\item{level}{of standard errors}

\item{stratas}{wich strata to plot (number or vector between 0 and nstrata-1)}

\item{relsurv}{set to FALSE. If TRUE relative survival curves are plotted.}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Plotting the cumulative baselines excess hazard
}
\examples{
excplot(m, se=TRUE, col=c("green","blue")) #with condifence intervals
excplot(m, se=TRUE, time=seq(0,30,1)) #at specific time-points
excplot(m, se=TRUE, relsurv=TRUE) #plot the relative survival
excplot(m, se=TRUE, polygon=FALSE) #No polygon
excplot(m, se=FALSE) #no confidence interval
excplot(m, se=TRUE, stratas=1) #plot the second strata
}
\author{
Cristina
}
