% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{matchpropexc}
\alias{matchpropexc}
\title{Excess risk paired survival model}
\usage{
matchpropexc(formula, data, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph}); use strata() for strata-variables}

\item{data}{data frame - already set-up. (see\code{compdata} for modre details)}

\item{...}{Additional arguments to lower level funtions}
}
\value{
no output. use \code{summary(model)} to view the coefficient estimates.
}
\description{
Excess risk paired survival model
}
\examples{
dhaz<-sim.data.MatchH(nca=5000, ncont=5)
setdhaz<-compdata(Surv(time, status)~x+z+cc, clust=id, idControl=j, data=dhaz)
names(setdhaz) # it is strongly recommended to check the 
               #names of your variables before estimating the model
exc.model<-matchpropexc(Surv(entry,exit,status)~strata(z)+factor(x),
                        data=setdhaz)
summary(exc.model)
### with competing risks
dhazc<-sim.data.MatchH(nca=5000,ncont=5,competing=TRUE)
setdhaz1<-compdata(Surv(time, status==1)~x+z+cc, clust=id, idControl=j, data=dhazc)
exc.model1<-matchpropexc(Surv(entry,exit,status)~strata(z)+factor(x)+cc, data=setdhaz1)
summary(exc.model1)
setdhaz2<-compdata(Surv(time, status==2)~x+z+cc, clust=id, idControl=j, data=dhazc)
exc.model2<-matchpropexc(Surv(entry,exit,status)~1, data=setdhaz2)
summary(exc.model2)
}
\author{
Cristina Boschini
}
