% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{summary.matchpropexc}
\alias{summary.matchpropexc}
\title{model summary}
\usage{
\method{summary}{matchpropexc}(object, ...)
}
\arguments{
\item{object}{model estimated with matchpropexc}

\item{...}{Additional arguments to lower level funtions}
}
\value{
coefficient estimates, standard error and significance level
}
\description{
model summary
}
\author{
Cristina Boschini
}
