% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{compdata}
\alias{compdata}
\title{Data structured for matchpropexc}
\usage{
compdata(formula, data, cluster, idControl, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph}); avoid using strata() or factor()}

\item{data}{data frame}

\item{cluster}{vector cluster indicator (one cluster for each exposed individual)}

\item{idControl}{vector control indicator (idControl==1 indicates exposed individual in cluster i)}
}
\description{
Data structured for matchpropexc
}
\examples{
dd<-data.sim(nca=5000, ncont=5)
setdd<-compdata(Surv(time, status)~x+z+cc, cluster=id, idControl=j, data=dd)
summary(setdd)
}
\author{
Cristina Boschini
}
