% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSim.R
\name{data.sim}
\alias{data.sim}
\title{Excess risk paired survival model}
\usage{
data.sim(nca, ncont, competing = FALSE, nullmod = FALSE)
}
\arguments{
\item{nca}{number of exposed individuals}

\item{ncont}{number of unexposed individuals for eac exposed (fixed number)}

\item{competing}{if TRUE a competing cause is considered}

\item{nullmod}{if TRUE no covariates are simulated}
}
\description{
Excess risk paired survival model
}
\author{
Cristina Boschini
}
