% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchcif.R
\name{ecif.coef}
\alias{ecif.coef}
\title{Excess CIF prediction based on newdata and geepack::geese estimates.}
\usage{
ecif.coef(model, times, link = "log", level = 0.95)
}
\arguments{
\item{model}{geese object. To be defined if coefs and vcov are null.}

\item{times}{vector of timepoints as the one used to estimate the GEE model}

\item{link}{link used to estimate the model. Choose between c("log","logit","identity")}
}
\value{
table with coefficient estimates, sandwich standard error, function of coefficient for interpretation and p-value.
}
\description{
Excess CIF prediction based on newdata and geepack::geese estimates.
}
\examples{
dcif<-sim.data.MatchCR(1000,5)
tp<-c(0.5,1,2,5,10,15,25)
setdcif1<-compcomp(timereg::Event(time=FALSE,time2=time,cause=cause)~X1+X2, 
                   data=dcif, cluster=i, idControl=j, time.points=tp,
                   cens.formula=NULL, event=1)
exc.cif.mod1<-geepack::geese(Rt~-1+factor(h)+X1+X2,
                             data=setdcif1,
                             family="gaussian", #error distribution
                             mean.link = "log", #link function for Rt
                             corstr="independence", #correlation structure
                             id=clust.num, #cluster vector
                             weights=weights #censoring weights 
                             )
ecif.coef(exc.cif.mod1,times = tp, link = "log")
}
