% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchcif.R
\name{eCIFpredict}
\alias{eCIFpredict}
\title{Excess CIF prediction based on newdata and geepack::geese estimates.}
\usage{
eCIFpredict(model = NULL, times, formula, dataset,
  strata.levels = NULL, coefs = NULL, vcov = NULL)
}
\arguments{
\item{model}{geese object. To be defined if coefs and vcov are null.}

\item{times}{vector of timepoints as the one used to estimate the GEE model}

\item{formula}{model formula}

\item{dataset}{new data}

\item{strata.levels}{if CIF predicted for different strata, define strata levels}

\item{coefs}{coefficient estimates (model$beta). To be specified if model is NULL}

\item{vcov}{coefficient variance and covariance matrix (model$vbeta). To be specified if model is NULL, together with coefs.}
}
\value{
dataset with predicted values; ready to be used with ggplot2
}
\description{
Excess CIF prediction based on newdata and geepack::geese estimates.
}
