% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{excplot}
\alias{excplot}
\title{Plotting the cumulative baseline excess hazard}
\usage{
excplot(x, se = FALSE, time = NULL, add = FALSE, ylim = NULL,
  lty = NULL, col = NULL, legend = TRUE,
  ylab = "Baseline cumulative excess hazard", polygon = TRUE,
  level = 0.95, stratas = NULL, relsurv = FALSE, hline = TRUE, ...)
}
\arguments{
\item{x}{matchpropexc object}

\item{se}{to include standard errors}

\item{time}{to plot for specific time variables}

\item{add}{to add to previous plot}

\item{ylim}{to give ylim}

\item{lty}{to specify lty of components}

\item{col}{to specify col of components}

\item{legend}{to specify col of components}

\item{ylab}{to specify ylab}

\item{polygon}{to get standard error in shaded form}

\item{level}{of standard errors}

\item{stratas}{wich strata to plot (number or vector between 0 and nstrata-1)}

\item{relsurv}{set to FALSE. If TRUE relative survival curves are plotted.}

\item{hline}{set to TRUE. An horizontal line is plotted (1 if relative survival, 0 if excess risk)}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Plotting the cumulative baselines excess hazard
}
\examples{
dd<-data.sim(nca=5000, ncont=5)
setdd<-compdata(Surv(time, status)~x+z+cc, data=dd, idControl = j, clust=id)
m <- matchpropexc(Surv(exit,status)~strata(z)+factor(x),data=setdd)
excplot(m, se=TRUE, col=c("green","blue"), main="with polygon CI") 
excplot(m, se=TRUE, time=seq(0,30,1), main="at specific time-points") 
excplot(m, se=TRUE, relsurv=TRUE, main="relative surv.") 
excplot(m, se=TRUE, polygon=FALSE, main="with CI - no polygon") 
excplot(m, se=FALSE, main="No CI") 
excplot(m, se=TRUE, stratas=1, main="plot the second strata")
}
\author{
Cristina
}
