% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{matchpropexc}
\alias{matchpropexc}
\title{Excess risk paired survival model}
\usage{
matchpropexc(formula, data, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph}); use strata() for strata-variables}

\item{data}{data frame - already set-up. (see\code{compdata} for modre details)}

\item{...}{Additional arguments to lower level funtions}
}
\value{
no output. use \code{summary(model)} to view the coefficient estimates.
}
\description{
Excess risk paired survival model
}
\examples{
dd<-data.sim(nca=5000, ncont=5)
setdd<-compdata(Surv(time, status)~x+z+cc, clust=id, idControl=j, data=dd)
names(setdd) #it is strongly recommended to check the names of your variables before estimating the model
exc.model<-matchpropexc(Surv(entry,exit,status)~strata(z)+factor(x), data=setdd)
summary(exc.model)
exc.model1<-matchpropexc(Surv(entry,exit,status)~1, data=setdd)
summary(exc.model1)
}
\author{
Cristina Boschini
}
