% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchsurv-package.R
\docType{package}
\name{matchsurv-package}
\alias{matchsurv-package}
\title{Excess Risk Estimation for Matched Cohort Survival Data}
\description{
tools for estimating the excess risk of exposed subjects when matched survival data are available.
}
\examples{

## To appear

}
\author{
Cristina Boschini
}
\keyword{package}
