% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimCIF.R
\name{sim.data.MatchCR}
\alias{sim.data.MatchCR}
\title{Function to simulate matchced survival data in competing risk setting based on cumulative incidence functions}
\usage{
sim.data.MatchCR(nca, ncont = 5, cifs = NULL, gammax = c(0.1, -0.2),
  mean.link = "log", bias = FALSE)
}
\arguments{
\item{nca}{number of exposed individuals}

\item{ncont}{number of unexposed individuals for each exposed (fixed number; by default ncont=5)}

\item{cifs}{list argument. Starting CIFs for: 1) 1st cause background risk, 2) 2nd cause background risk, 3) 1st cause excess risk, 4) 2nd cause excess risk. If cifs=NULL default quantities are used.}

\item{gammax}{coefficients for the excess risk model. Max 2 values to be specified (the first covariate is by default binomial, the second is log normal), Default values: $\gamma_1=0.1$, $\gamma_2=-0.2$. Note: some value might not work.}

\item{mean.link}{link function for the gee model. The data are simulating accordingly to the link that will be used to estimate the excess cif model. Two values supported "id" or "log".}

\item{bias}{to simulate data with excess risk factor correlated with age}
}
\value{
The function returns a dataset
}
\description{
Function to simulate matchced survival data in competing risk setting based on cumulative incidence functions
}
\author{
Cristina Boschini
}
