% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{summary.matchpropexc}
\alias{summary.matchpropexc}
\title{model summary (coefficient estimates, SE estimates and significance level)}
\usage{
\method{summary}{matchpropexc}(object, ...)
}
\arguments{
\item{object}{model estimated with matchpropexc}
}
\description{
model summary (coefficient estimates, SE estimates and significance level)
}
\author{
Cristina Boschini
}
