% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchf.R
\name{vcov.matchpropexc}
\alias{vcov.matchpropexc}
\title{variance and covariance matrix for the coefficient estimates}
\usage{
\method{vcov}{matchpropexc}(object, ...)
}
\arguments{
\item{object}{model estimated with matchpropexc}
}
\value{
variance and covariance matrix of the coefficient estimates
}
\description{
variance and covariance matrix for the coefficient estimates
}
\author{
Cristina Boschini
}
